[Scala] Use distinct scripts to run & test the sistem
This fixes the fact that two test classes may cause the system to start two different JVMs using the same configuration file.

If this happens and Akka.Remote is enabled, this will likely cause the test to crash since other JVMs than the first to start will find the address already in use.